Adds stylesheet component for items
Item components are used to mimic the appearance from forms when the
content isn't actually editable but is still arranged in the same way as
the form is.

For example, you might have a widgets/edit view and a widgets/show view;
you might want the show view to look like the edit view, except that the
values would be displayed rather than in form fields.